Spec maseya_blend
While the specific color adjustments are arbitrary in a general sense and evolving, for each given commit they are a fact of the implementation and we can spec for those given specifics.

Due to how luma adjustments can clamp color components, some tests can only reliably verify a subset of the hcy space. We are making the assumption the rest of the input domain is calculated correctly.