Switch to a generator function for blend values
A default generator function is used if none is passed.

For most modes, the triple of values used to vary colors are not used as a color per se, so we will refer to them as blend values.

The PRNG is also extended to allow for a seed value. Since the JS standard `Math.random` method does not allow for seeding we use a new implementation. A JS port of the sfc32 algorithm (https://sourceforge.net/projects/pracrand/) is used as the PRNG core as it is fast while also demonstrating sufficient random distribution. It can be seeded with 1, 2, or 3 values in the 32 bit range. If no seed values are passed `Math.random` is used to generate three 32 bit seed values.